Added Retrieve Command Also refactored commands to use currying to prevent code duplication, as most Get requests should be relatively similar.
