+++ src/main/scala/de/upb/cs/swt/delphi/cli/DelphiCLI.scala
@@ -1,5 +1,7 @@
 package de.upb.cs.swt.delphi.cli
 
+import de.upb.cs.swt.delphi.cli.commands.{RetrieveCommand, TestCommand}
+
 
 /**
   * The application class for the Delphi command line interface
@@ -20,8 +22,14 @@ object DelphiCLI extends App {
       checkConfig(c => if (c.server.isEmpty()) failure("Option server is required.") else success)
 
       cmd("test").action((_,c) => c.copy(mode = "test"))
+
+      cmd("retrieve").action((s,c) => c.copy(mode = "retrieve"))
+        .text("Retrieve a project's description, specified by ID.")
+        .children(
+          arg[String]("ID").action((x, c) => c.copy(args = List(x))).text("The ID of the project to retrieve")
+        )
+
       //cmd("search")
-      //cmd("retrieve")
     }
   }
 
@@ -31,6 +39,7 @@ object DelphiCLI extends App {
       cliParser.showHeader()
       config.mode match {
         case "test" => TestCommand.execute(config)
+        case "retrieve" => RetrieveCommand.execute(config)
         case _ => println("Unknown command")
       }
